requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
motor>=3.1.1
beanie>=1.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
PyJWT>=2.8.0
python-dotenv>=0.19.0
email-validator>=2.0.0 